+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All Notable changes to `Laravel Follow` will be documented in this file
 
+### v2.0.0 (2018-03-06)
+
+`ADDED`
+
+- Support for Laravel 5.6
+
+### v1.1.0 (2018-03-05)
+
+`ADDED`
+
+- `followMany()` method on `CanFollow` trait to follow many entities at once.
+- `unfollowMany()` method on `CanFollow` trait to unfollow many entities at once.
+- `syncManyFollowings()` method on `CanFollow` trait to sync many followings at once.
+- `addManyFollowers()` method on `CanBeFollowed` trait to add many followers at once.
+- `removeManyFollowers()` method on `CanBeFollowed` trait to remove many followers at once.
+- `syncManyFollowers()` method on `CanBeFollowed` trait to sync many followers at once.
+
 ### v1.0.1 (2017-10-26)
 
 `ADDED`